removed linear ramp up, as it can be achieved with corresponding paramters with the logistic curve aswell. Both have a part with a fast build up of speed, but for few steps with precision logistic curve is better. The current settings of motor speed are appropriate when using the motor with an arm. Not to fast, but not to slow either.
When rotating the arm the release signal must be set with care, as high precision movement is impossible, when releasing the arm. Those cables are holding considerable tension, which let the motor moving in the opposite direction as intented due to the movement after releasing the arm due to the cable tension.

Heat production when holding the arm steady should be monitored, but as of now, it does not seem to hold a problem.