The Lepton Software Interface document defines a function `SYS Number of Frames to Average`. This function can be used to set another amount of frames to average. The purethermal breakout board sets different values, but the shutter for Lepton 3.5 does not obey those values resulting in no quality improvement of the FFC
After restart of the Lepton module the temperature readings seem to be most accurate (Skin temperature about 36-37°C). After 1 minute without FFC the accuracy diminishes to 33-34°C. When FFC is issued, the readings stay at about 35°C for approx. 10 seconds and then diminish again to 33-34°C.

The 35°C readings after FFC are reproducible with an error about 0.5°C of the highest reading. I propose for this ticket to do a manual FFC right before taking the IR-image at the desired position and then taking the image with the highest temperatures in a 5 second image taking intervall.

After FFC the temperature readings for skin are about 34°C and then rise to 35°C over the next couple of seconds, before falling again. That is the rational behind taking the image with the highest temperature readings.

This ticket only refactors ffc running into another function. The function should be callable in the future by the client via signals, instead of being issued every 100 frames.